Changed how database is accessed
Just learned that MongoClient can be reopened once closed, which means I could use one instance for the lifetime of the application. Did some reading and it's actually recommended as well. Now there is only one instance of our MongoClient connection to the database, and it is closed and reopened when it needs to be. Other modules still use get_db() to get the instance because I feel it is cleaner that way, instead of just directly accessing the variable