Improve calendarview/index.html Jinja templating
Readability of the logic used in calendarview/index.html is a bit
improved.

The way in which a calendar is decided to be shown is slightly changed as well.
Depending on the 'calendar' variable passed in to the template, the id of the
calendar div is changed to the calendar that should be shown. This is done
by the new getConnectedCalendar() function in calendar.js.