Fix some mistakes from commit 37e92
In calendar.js, the id attribute taken from the calendar wrapper was being
compared to "Google" instead of "google". I think I accidently deleted
all of the constants, but they're back now.

In calendarAuth.js, event listeners are added to buttons that may or may not
exist depending on the user. This caused an error if they didn't exist, so
there are now appropriate checks to see if buttons exist in the first place.