JS scripts loaded in calendar page use 'defer' attribute
Was having some problems with content on the DOM not being available
to me on time, and didn't really like putting adding event listeners
in a DOMContentLoaded handler. Adding defer to loaded scripts does
exactly what I need. Make sure the script is run when everything on
the DOM has been parsed.