Finish delete shift functionality
In order to delete a Shift, a user has to open the dropdown menu
and click the trash icon next to the Shift. This was a little bugged
before this commit, but now it's fixed and works as it's supposed
to. It's not completely finished, as I'm about to show below.

api.py now responds with the name of the Shift created, which makes it
easier for shifts.js to add it to the dropdown menu upon creation.

TODO: The shift does not disappear immediately, so will have to add
some deleting animation to let user know it is in the process of being
deleted.