Add animation when deleting a Shift
When you click the trash icon to delete a Shift, it does not disappear
immediately, since it's not actually deleted immediately. In order to
let the user know that their delete request was acknowledged, a pulsing
animation is added to the Shift in question. If for whatever reason, an
error occurs when deleting a Shift, the animation is removed and the Shift
is not removed.