Add hover animations for Shift selection dropdown menu
The dropdown menu opener changes the cursor to a pointer to better
show that it should be clicked. The options themselves do not have a
pointer cursor, but they now become slightly darker when hovered over.

The trash icon now has the same background color as its container, and
is black when not being hovered on. It changes to red when it is being
hovered on, which I think makes things clearer.