Update underlying <select> when updating custom select (dropdown menu)
When a new option is selected in the dropdown menu, the underlying <select>
wasn't updated to reflect that. It does now. Also changed the braindead way
I was unhighlighing the previous selected option.

Also added 'click' event listener to each date on the calendar to prepare on
working on actually adding a Shift to a user's calendar.