Add 'pending' shifts when clicking date on calendar
When a date on the calendar is clicked, the selected shift is added to the
box and will have a small animation indicating that it is 'pending'. Pending
meaning that it hasn't been added to the actual calendar, that will be done
by some button that appears when there are pending shifts. This button hasn't
been made yet.

The name of the current month is now abbreviated on the calendar. My conversion
table for months to their number uses the abbreviated version, so this makes
things easier.