Fix a couple mistakes and bugs
1. The event listeners added to the date boxes were being added before the
date boxes were created, which meant clicking the date boxes had no effect.
Removed adding the event listeners, and I'll have to place them somewhere
when the date boxes are guaranteed to be made.

2. Fixed the messed up placement of arguments when creating the default
'Select a Shift:' option. Now the value attribute of that option is '-1',
instead of the 'Select a Shift:' it mistakenly was before.

3. Removed the redundant '.' in front of a class name that I was deleting
using elem.classList.remove().

4. When checking the selected shift when adding a pending shift to the
calendar, I mistakenly used the integer -1 when checking for the default
'Select a Shift:' option. Changed to the string '-1'