Turn scripts.js and calendar.js into modules and A TON of bug fixes                                                                            | I just kept running into bug after bug, so this commit is a big mess, but it's not like there aren't other commits like this anyway.
1. Decided that calendar.js would be responsible for adding Shifts to
calendar, but it would need some functions that I felt belonged to
shifts.js, so I turned both scripts to modules and exported the relevant
functions from shifts.js. These are getSelectedShift() and concatDateShift().

2. Organized how the calendar is created a bit more, which I think makes it
a bit more readable, and increases the consistency in using the appendChild()
methods to manipulate the DOM. Allows me to add the click event listener to
each date box as they're created, fixing an earlier issue.

3. I never noticed that the 'Go back' button in the Create Shift view broke
when I had the scripts deferred. The event listener is now added in shifts.js.

4. Don't know how the highlighting the selected Shift broke, but it did, and
it's fixed now.