Populate config section of Calendar View page
I plan on having a <select> (dropdown menu) as one of the components
on the config section. It would allow the user to switch between
connected calendars if they had more than one.

In the middle I will place the buttons that allow for connecting
to different calendar vendors.

The right portion of the config section currently holds the 'Confirm'
button for confirming pending shifts on the calendar and adding them
to the underlying calendar. It already has very limited functionality,
which is being disabled when there are no pending shifts.