List all calendars connected to a single Google account
Google accounts (and other calendar vendors) allow forr a user to have
multiple calendars in a single account. This commit sets up the
foundations for users to be able to add Shifts to all their calendars.

For now, the calendars are just listed in a <select> element in the
config section of the Calendar View (above the calendar).  This will
be changed to a custom select menu that the user will use to switch
between calendars.