Allow for adding Shifts to all calendars (Google)
Some calendar vendors, like Google, let a user have multiple different
calendars on a single account.

I have added a 'Switch Calendars' dropdown menu on top of the calendar
will let the user switch between the calendars on their account. The menu
is created in customElems.js and the GET request for the calendars is
done in calendar.js.

Also added a couple more variables in calendar.js (calendarVendor) to
isolate different calendar vendors, in order to make it easier in the
future when I add more vendors (eg. Microsoft) to make the correct
 API calls to the correct vendors.

 I probably didn't get everything. I need to stop these stupid big commits.