Stop cursor from changing in 'Select a Shift' dropdown menu
The options in the 'Select a Shift' dropdown menu have text (the name
of the Shift) which caused the cursor to change to the I shaped cursor.
The cursor now stays as the default cursor. The element to open the
dropdown menu will continue to have the pointer cursor though, I
think that makes things more clear.

The 'Switch Calendar' dropdown menu already did this.