Fix value attribute of calendar <option>s
The hidden <select> element that underlies the 'Switch Calendars'
dropdown menu has <option>s with the id of each calendar as
their value. Issue was that all the <option>s aside from the primary
calendar <option> had a value of undefined instead of the calendar's
id.

That is fixed now.