Google calendar users can now add Shifts to their calendar. (plus a ton of api changes)
Future Maxwell, I promise that this will be the last big commit I do (hopefully).

- Users can now add Shifts to their Google calendars, All calendars that
they have write access to. In the future, they won't be able to even see
calendars that they don't have write access to.

- Changed how requests to external api's (eg. Google Calendar API, etc)
are made to make it easier to handle external api errors.

The actual request is made inside an inner function, which
returns the response. If the response returns a bad status code,
the inner function that makes the request is passed to an error
error handler that will handle the bad response and use the
function to retry the request. It will return the response if the
request succeeds, or None if it retries the request multiple times
and can't get a successful response.

This is not fully implemented and just the barebones are written
right now.