Use 'fields' parameter to reduce respone length
The Google Calendar API accepts a 'fields' query argument that can
be used to filter out fields in the response you don't need. Given that
I don't need most of the fields included in the responses I get in
GoogleAuth.list_events() and GoogleAuth.list_calendars(), I added this
'fields' param to the request. It should improve performance as well.