Allow user to switch between months quickly
Shrinking the size of the spinner container allowed me to click the
next/previous month buttons quicker, but the events could get
messed up and the spinner animation was erratic.

I added a new variable that keeps track of the current drawEvents()
request and is able to abort the request if the user switches to another
month quickly.