Fix bug when confirming Shifts to the calendar
A bug was introduced when I added the next & previous month buttons
where the month is written on the calendar. I use that element to figure
out what month the user is viewing, and adding the buttons messed
things up when it came to that, which caused in a 'Bad Request'
response from Google Calendar.

Fix was adding the name of the month to the value attribute of that
element, which i think is pretty semantic. Might add the year the user
is viewing in the future as well.