Give names to calendars that the user connects
Here comes another huge commit.

This program already allows the user to switch between calendars
in the account they connected with (currently only Google). The thing
is that I also want this application to allow the user to connect to
several Google (or other vendors) accounts and access those calendars
as well, and to do that was a bigger challenge than I expected.

First thing is that I had to change the layout of a user document in
the MongoDB database. It now supports a user to have multiple
calendar accounts and better handles how to link a certain
calendar with the credentials needed to manipulate it (tokens
and the such).

Next was removing the access token from the session cookie, which
is something I shouldn't have done in the first place. Because of
the better placing of credentials, I get the access_token for a
particular calendar from the database.

Session cookie now stores the current calendar, and by knowing which
vendor the current calendar is from, the client can send the right
requests for the right vendors. Thinking of it now, sending the right
requests should maybe be handled server side, but we'll see.

Biggest point is that the user must give a name to the calendar account
they're about to connect, which will allow me to let user switch between
not only the calendars in a single account, but calendars between
several accounts.

Thank god I'm probably the only one ever going to work on this repo, I
wouldn't wish maintaining this on anyone...