User should be able to switch between calendar accounts
I haven't tested this but they should be able to now.

I have to say that the way I implemented this was weird. It changes
which calendar account is shown by reloading the page and putting
the name and vendor of the calendar in the query string, where
calendar.py index() handles everything else.