Re-organize where some code is placed
The oauth.py module will hold the endpoints used to connect a new calendar and handle
the OAuth workflow for different vendors (only Google rn). The
oauth_calendar_classes.py will holds the classes that actually make calls to
the specific API that will make changes to the user's calendar.